feat(seo): server-render homepage hero to fix zero-word crawler output
The homepage (/en) was returning 0 server-rendered words to search and
AI crawlers while /en/privacy returned ~493 words. Root causes:

1. page.tsx was "use client" and contained Math.random() in a useState()
   initializer, causing a hydration mismatch that made Next.js bail on
   SSR and produce an empty shell for the client component subtree.

2. The welcome heading/description were only rendered client-side, so
   Google and AI bots saw no body text.

Fix:
- Extract the client interactive logic into ChatIsland.tsx (keeps "use
  client"); accepts an optional heroContent prop for server-rendered slots.
- Replace page.tsx with a thin server component that calls
  getTranslations() and passes the hero h2+p as a React node into
  ChatIsland. The composition pattern guarantees the text is in the
  initial SSR HTML regardless of what the client island does.
- Move the suggestedPrompts Math.random() shuffle from the useState()
  initializer (runs on both server and client, causing a mismatch) into
  a useEffect() (client-only), so server and first client render agree.
- Update page.test.tsx import to point to ChatIsland (all 591 tests pass).
